REF: Archetypes with SoA
Performance boot up to 30% (hopefully)

+++ script/script.cpp
@@ -189,7 +189,7 @@ namespace script
     auto res = scriptECS->dataQueries.find(type->GetSubTypeId());
     assert(res != scriptECS->dataQueries.end());
 
-    return new (RawAllocator<int>::alloc()) size_t(res->second.entitiesCount);
+    return new (RawAllocator<int>::alloc()) int(res->second.entitiesCount);
   }
 
   int create_script_query_count_get(size_t *sz)
